+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning].
 - BREAKING: Remove branch configuration option for scheduled runs.
 - BREAKING: Remove option to commit changes.
 - BREAKING: Remove `fileCount` as templating variable.
+- BREAKING: Disabling from push event's now disables the action entirely.
 
 ## [1.2.2] - 2021-02-08